Rename Projects Page to ProjectsPage
